refactor(cli): decompose ops/bench.rs into a bench/ module tree
Splits the ~2200-line bench harness into bench/{mod,stats,report,
regression,harness,probes}.rs by concern: timing/stats, report schema
+ persistence, baseline regression gate, probe-generic orchestration,
and the libkrun/HVF/Firecracker probe impls. Pure structural move:
mvmctl bench output and every JSON report schema are byte-identical,
and bench::Args / bench::run / BootMarks / the boot-timing-sidecar
writer still resolve at the same paths for group.rs and bench_probe.rs.
Updates the check-cli-runtime-surface exemption to the file that now
carries the AnyBackend dispatch.
