+++ xtask/src/check_cli_runtime_surface.rs
@@ -86,7 +86,7 @@ const EXEMPTIONS: &[(&str, &[Rule], &str)] = &[
         "pool build substrate: backend selection",
     ),
     (
-        "commands/ops/bench.rs",
+        "commands/ops/bench/probes.rs",
         &[Rule::AnyBackend],
         "benchmarking substrate: backend selection",
     ),
